Splitting the scraping codes into files
