+++ .gitignore
@@ -5,6 +5,8 @@ __pycache__/
 
 # C extensions
 *.so
+.compras
+
 
 # Distribution / packaging
 .Python
