+++ requirements-test.txt
@@ -1,3 +1,4 @@
+-r requirements.txt
 pytest==3.5.1
 pytest-cov==2.5.1
